rename frozen_attribute to attribute_helper
