+++ core/search/all_iterator.hpp
@@ -29,7 +29,7 @@
 #include "search/cost.hpp"
 #include "search/score.hpp"
 #include "search/sort.hpp"
-#include "utils/frozen_attributes.hpp"
+#include "utils/attribute_helper.hpp"
 
 namespace iresearch {
